Move findCreatorAnnotation
to reduce overhead
