+++ src/main/kotlin/com/fasterxml/jackson/module/kotlin/annotation_introspector/KotlinPrimaryAnnotationIntrospector.kt
@@ -1,9 +1,12 @@
 package com.fasterxml.jackson.module.kotlin.annotation_introspector
 
+import com.fasterxml.jackson.annotation.JsonCreator
 import com.fasterxml.jackson.annotation.JsonProperty
 import com.fasterxml.jackson.databind.DeserializationFeature
 import com.fasterxml.jackson.databind.Module
+import com.fasterxml.jackson.databind.cfg.MapperConfig
 import com.fasterxml.jackson.databind.introspect.Annotated
+import com.fasterxml.jackson.databind.introspect.AnnotatedConstructor
 import com.fasterxml.jackson.databind.introspect.AnnotatedField
 import com.fasterxml.jackson.databind.introspect.AnnotatedMember
 import com.fasterxml.jackson.databind.introspect.AnnotatedMethod
@@ -17,13 +20,17 @@ import com.fasterxml.jackson.module.kotlin.reconstructClass
 import com.fasterxml.jackson.module.kotlin.toSignature
 import kotlinx.metadata.Flag
 import kotlinx.metadata.KmClass
+import kotlinx.metadata.KmClassifier
 import kotlinx.metadata.KmProperty
+import kotlinx.metadata.KmValueParameter
 import kotlinx.metadata.jvm.fieldSignature
 import kotlinx.metadata.jvm.getterSignature
 import kotlinx.metadata.jvm.setterSignature
 import kotlinx.metadata.jvm.signature
 import java.lang.reflect.AccessibleObject
+import java.lang.reflect.AnnotatedElement
 import java.lang.reflect.Constructor
+import java.lang.reflect.Executable
 import java.lang.reflect.Method
 
 // AnnotationIntrospector that overrides the behavior of the default AnnotationIntrospector
@@ -132,4 +139,63 @@ internal class KotlinPrimaryAnnotationIntrospector(
 
         return requiredAnnotationOrNullability(byAnnotation, byNullability)
     }
+
+    // Return Mode.DEFAULT if ann is a Primary Constructor and the condition is satisfied.
+    // Currently, there is no way to define the priority of a Creator,
+    // so the presence or absence of a JsonCreator is included in the decision.
+    // The reason for overriding the JacksonAnnotationIntrospector is to reduce overhead.
+    // In rare cases, a problem may occur,
+    // but it is assumed that the problem can be solved by adjusting the order of module registration.
+    override fun findCreatorAnnotation(config: MapperConfig<*>, ann: Annotated): JsonCreator.Mode? {
+        (ann as? AnnotatedConstructor)?.takeIf { 0 < it.parameterCount } ?: return null
+
+        val declaringClass = ann.declaringClass
+        val kmClass = declaringClass
+            ?.takeIf { !it.isEnum }
+            ?.let { cache.getKmClass(it) }
+            ?: return null
+
+        return JsonCreator.Mode.DEFAULT
+            .takeIf { ann.annotated.isPrimarilyConstructorOf(kmClass) && !hasCreator(declaringClass, kmClass) }
+    }
+}
+
+private fun Constructor<*>.isPrimarilyConstructorOf(kmClass: KmClass): Boolean = kmClass.findKmConstructor(this)
+    ?.let { !Flag.Constructor.IS_SECONDARY(it.flags) || kmClass.constructors.size == 1 }
+    ?: false
+
+private fun AnnotatedElement.hasCreatorAnnotation(): Boolean =
+    annotations.any { it is JsonCreator && it.mode != JsonCreator.Mode.DISABLED }
+
+private fun KmClassifier.isString(): Boolean = this is KmClassifier.Class && this.name == "kotlin/String"
+
+private fun isPossibleSingleString(
+    kotlinParams: List<KmValueParameter>,
+    javaFunction: Executable,
+    propertyNames: Set<String>
+): Boolean = kotlinParams.size == 1 &&
+    kotlinParams[0].let { it.name !in propertyNames && it.type.classifier.isString() } &&
+    javaFunction.parameters[0].annotations.none { it is JsonProperty }
+
+private fun hasCreatorConstructor(clazz: Class<*>, kmClass: KmClass, propertyNames: Set<String>): Boolean {
+    val kmConstructorMap = kmClass.constructors.associateBy { it.signature?.desc }
+
+    return clazz.constructors.any { constructor ->
+        val kmConstructor = kmConstructorMap[constructor.toSignature().desc] ?: return@any false
+
+        !isPossibleSingleString(kmConstructor.valueParameters, constructor, propertyNames) &&
+            constructor.hasCreatorAnnotation()
+    }
+}
+
+// In the original, `isPossibleSingleString` comparison was disabled,
+// and if enabled, the behavior would have changed, so the comparison is skipped.
+private fun hasCreatorFunction(clazz: Class<*>, kmClass: KmClass): Boolean = kmClass.companionObject
+    ?.let { companion ->
+        clazz.getDeclaredField(companion).type.declaredMethods.any { it.hasCreatorAnnotation() }
+    } ?: false
+
+private fun hasCreator(clazz: Class<*>, kmClass: KmClass): Boolean {
+    val propertyNames = kmClass.properties.map { it.name }.toSet()
+    return hasCreatorConstructor(clazz, kmClass, propertyNames) || hasCreatorFunction(clazz, kmClass)
 }
